Move api/server/router/plugin to daemon/server/router/plugin
